\name{dinvgauss}
\alias{dinvgauss}
\title{Inverse Gaussian distribution - Density.}
\description{Computes de p.d.f. function for the inverse Gaussian distribution.}
\usage{dinvgauss(y, m, s, logcond=FALSE)}
\arguments{
		\item{y}{\eqn{p}-dimensional vector with positive entries}
		\item{m}{Mean parameter}
		\item{s}{Scale parameter}
		\item{logcond}{Logic parameter for log-scale form}
        }        
\references{
	-	"Utilities for Nonlinear Regression and Repeated Measurements Models." Package "rmutil" version 1.0.
			}
\author{
	Jim Lindsey <jlindsey@luc.ac.be>
       }