\name{pinvgauss}
\alias{pinvgauss}
\title{Inverse Gaussian distribution - Probability.}
\description{Computes de c.d.f. function for the inverse Gaussian distribution.}
\usage{pinvgauss(q, m, s)}
\arguments{
		\item{q}{\eqn{p}-dimensional vector with (0,1)-valued entries}
		\item{m}{Mean parameter}
		\item{s}{Scale parameter}
        }        
\references{
	-	"Utilities for Nonlinear Regression and Repeated Measurements Models." Package "rmutil" version 1.0.
			}
\author{
	Jim Lindsey <jlindsey@luc.ac.be>
       }