\name{popcomp}
\alias{popcomp}
\docType{data}
\title{Composition of the simulated population segemented in planned and unplanned domains.}
\description{This object corresponds to a list with the composition of the simulated population across planned and unplanned domains. See Martinez-Ovando et al. (2014) for further details.}
\usage{data(popcomp)}
\format{
  The format is:
List of 3
 $ : num [1:2] 800 1200
 $ : num [1:3] 980 800 220
}
\source{Simulated data. See Martinez-Ovando et al. (2014) for further details.}
\references{
	- "Predictive inference on finite populations segmented in planned and unplanned domains," Martinez-Ovando, J. C., Olivares-Guzman, S. I., Roldan-Rodriguez, A., 2014. Submitted. Also available as Banco de Mexico Working Paper 2014-04. 	
}
\examples{data(popcomp)}
\keyword{datasets}