\name{poptot}
\alias{poptot}
\docType{data}
\title{Totals of the simulated population segemented in planned and unplanned domains.}
\description{
	This object corresponds to a list with the totals of the simulated population across 
	planned and unplanned domains. See Martinez-Ovando et al. (2014) for further details.
	}
\usage{
	data(poptot)
	}
\format{
	The format is:
	\describe{
		\item{[[1]]}{Composition of the population across planned domains (working format).}
		\item{[[2]]}{Composition of the population across planned domains.}
		\item{[[3]]}{Composition of the population across unplanned domains .}
		}
	}
\source{
	Simulated data. See Martinez-Ovando et al. (2014) for further details.
	}
\references{
	- "Predictive inference on finite populations segmented in planned and unplanned domains," Martinez-Ovando, J. C., Olivares-Guzman, S. I., Roldan-Rodriguez, A., 2014. Submitted. Also available as Banco de Mexico Working Paper 2014-04. 	
}
\examples{data(poptot)}
\keyword{datasets}