\name{predfinitepop}
\alias{predfinitepop}
\title{Predictive Inference on Totals and Averages of Finite Populations Segmented in Planned and Unplanned Domains.}
\description{This package computes Bayesian predictive inference on totals and averages of finite populations segmented in planned and unplanned domains. Inference is based on Bayesian nonparametric methods using species-sampling models (specifically the Dirichlet process).}
\details{
		\tabular{ll}{
					Package: \tab predfinitepop\cr
					Type: \tab Statistical\cr
					Version: \tab v-1\cr
					Date: \tab 2014-04-24\cr
					}
		}			
\author{
	\tabular{ll}{
		\tab Juan Carlos Martinez-Ovando\cr
		\tab Sergio I. Olivares-Guzman\cr
		\tab Adriana Roldan-Rodriguez\cr
		}
	}
\references{
    - "A Bayesian nonparametric framework to inference on totals of finite populations," Martinez-Ovando, J. C., Olivares-Guzman, S. I., Roldan-Rodriguez, A., 2013. Contributions of Young Researchers to Bayesian Statistics (eds. E. Lazarone & F. Ieva), Chapter 15 Springer.

	- "Predictive inference on finite populations segmented in planned and unplanned domains," Martinez-Ovando, J. C., Olivares-Guzman, S. I., Roldan-Rodriguez, A., 2014. Submitted. Also available as Banco de Mexico Working Paper 2014-04. 	
	   }