\name{qinvgauss}
\alias{qinvgauss}
\title{Inverse Gaussian distribution - Quantile.}
\description{Computes the quantiles for the inverse Gaussian distribution.}
\usage{qinvgauss(p, m, s)}
\arguments{
		\item{p}{\eqn{p}-dimensional vector with (0,1)-valued entries}
		\item{m}{Mean parameter}
		\item{s}{Scale parameter}
        }        
\references{
	-	"Utilities for Nonlinear Regression and Repeated Measurements Models." Package "rmutil" version 1.0.
			}
\author{
	Jim Lindsey <jlindsey@luc.ac.be>
       }