\name{rinvgauss}
\alias{rinvgauss}
\title{Inverse Gaussian distribution - Simulation.}
\description{Simulates reandom variables from the inverse Gaussian distribution.}
\usage{rinvgauss(n=1, m, s)}
\arguments{
		\item{n}{Number of simulatios}
		\item{m}{Mean parameter}
		\item{s}{Scale parameter}
        }        
\references{
	-	"Utilities for Nonlinear Regression and Repeated Measurements Models." Package "rmutil" version 1.0.
			}
\author{
	Jim Lindsey <jlindsey@luc.ac.be>
       }